refactor: mov the api function to a separate file
