+++ client/src/components/Seacrh/Seacrh.tsx
@@ -2,31 +2,31 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import './Seacrh.scss';
 import React, { useState, useEffect } from 'react';
 import { Container, Form, FormControl, Button } from 'react-bootstrap';
-import axios from 'axios';
+
 import Store from '../../store/Store';
-// import Api, { apiKey, apiFlickr } from '../../api/flickr';
+import Api from '../../api/flickr';
 
 const Seacrh = () => {
   const [inputText, setInputText] = useState<string>('');
   // const [images, setImages] = useState<object>({});
 
-  const apiKey = 'e538ac8584601a9e9bcd4c35b2315eaa';
+  // const apiKey = 'e538ac8584601a9e9bcd4c35b2315eaa';
 
-  const Api = {
-    getImages(textValue: string, page: string) {
-      return axios
-        .get(
-          `https://www.flickr.com/services/rest/?method=flickr.photos.search&api_key=${apiKey}&tags=${textValue}&per_page=10&page=${page}&format=json&nojsoncallback=1`
-        )
-        .then((response) => {
-          console.log(1);
-          return response.data.photos;
-        })
-        .catch((error) => {
-          console.log('Error fetching posts: ', error);
-        });
-    },
-  };
+  // const Api = {
+  //   getImages(textValue: string, page: string) {
+  //     return axios
+  //       .get(
+  //         `https://www.flickr.com/services/rest/?method=flickr.photos.search&api_key=${apiKey}&tags=${textValue}&per_page=10&page=${page}&format=json&nojsoncallback=1`
+  //       )
+  //       .then((response) => {
+  //         console.log(1);
+  //         return response.data.photos;
+  //       })
+  //       .catch((error) => {
+  //         console.log('Error fetching posts: ', error);
+  //       });
+  //   },
+  // };
 
   const changeHandle = (event: React.ChangeEvent<HTMLInputElement>) => {
     setInputText(event.target.value);
